requirements.txt
absl-py==2.0.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudpickle==3.0.0
comm==0.1.4
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.1
datadings==3.4.6
datasets==2.14.5
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@5de5319a49e7ea819586d3d6a817632392dfdeb2
dill==0.3.7
Django==4.2.6
django-cors-headers==4.2.0
executing==2.0.0
fastjsonschema==2.18.1
filelock==3.12.4
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
fvcore==0.1.5.post20221221
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.24.0
google-auth-oauthlib==1.1.0
greenlet==3.0.0
grpcio==1.59.3
huggingface-hub==0.18.0
hydra-core==1.3.2
idna==3.4
ijson==3.2.3
importlib-metadata==6.11.0
interchange==2021.0.4
iopath==0.1.9
ipykernel==6.25.2
ipython==8.16.1
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
langchain==0.0.306
langsmith==0.0.41
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.7
msgpack-numpy==0.4.8
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
natsort==8.4.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
neo4j==5.15.0
nest-asyncio==1.5.8
networkx==3.0
notebook==7.0.5
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.0
oauthlib==3.2.2
omegaconf==2.3.0
openai==0.28.1
opencv-python==4.9.0.80
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
pansi==2020.7.3
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.11.0
portalocker==2.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py2neo==2021.2.4
pyarrow==13.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
simplebloom==1.0.5
simplejpeg==1.7.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.3
streamlit==1.29.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.17.1
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
triton==2.1.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wikipedia==1.4.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.2
zipp==3.17.0
